* upstream/master: (674 commits)
  [Installer] Fix removeBinaries, closes #567
  Do not suggest packages that were installed
  Skip filter_var on 5.3.2 since it is buggy
  Do not prevent updates if untracked files are present
  Fix DownloadManagerTest creating directories on disk
  Improvements to autoload docs
  Add default http(s) ports to proxy configuration if they are missing.
  [docs] Change alias article according to comments
  [docs] Remove false information about recursive include fields
  [docs] Remove mentions of version_normalized
  [docs] Clarify that there is no cross-repo leakage with notify
  [docs] Correct default process timeout
  [docs] Improve description of COMPOSER_HOME
  [docs] Aliases: branch-alias and inline aliasing
  [docs] Document COMPOSER_HOME
  [docs] Update composer repository specification
  [docs] Document the notify-on-install config option
  Update lock
  Work around poorly configured php
  Add IOInterface::isDecorated()
  ...

Conflicts:
	src/Composer/Repository/Vcs/GitHubDriver.php
	src/Composer/Repository/Vcs/VcsDriver.php

+++ src/Composer/Util/Filesystem.php
@@ -10,9 +10,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Composer\Downloader\Util;
-
-use Composer\Util\ProcessExecutor;
+namespace Composer\Util;
 
 /**
  * @author Jordi Boggiano <j.boggiano@seld.be>
@@ -21,13 +19,22 @@ class Filesystem
 {
     public function removeDirectory($directory)
     {
+        if (!is_dir($directory)) {
+            return true;
+        }
+
         if (defined('PHP_WINDOWS_VERSION_BUILD')) {
             $cmd = sprintf('rmdir /S /Q %s', escapeshellarg(realpath($directory)));
         } else {
             $cmd = sprintf('rm -rf %s', escapeshellarg($directory));
         }
 
-        return $this->getProcess()->execute($cmd) === 0;
+        $result = $this->getProcess()->execute($cmd) === 0;
+
+        // clear stat cache because external processes aren't tracked by the php stat cache
+        clearstatcache();
+
+        return $result && !is_dir($directory);
     }
 
     public function ensureDirectoryExists($directory)
@@ -51,19 +58,25 @@ class Filesystem
      *
      * @param string $from
      * @param string $to
+     * @param Boolean $directories if true, the source/target are considered to be directories
      * @return string
      */
-    public function findShortestPath($from, $to)
+    public function findShortestPath($from, $to, $directories = false)
     {
         if (!$this->isAbsolutePath($from) || !$this->isAbsolutePath($to)) {
             throw new \InvalidArgumentException('from and to must be absolute paths');
         }
 
+        $from = lcfirst(rtrim(strtr($from, '\\', '/'), '/'));
+        $to = lcfirst(rtrim(strtr($to, '\\', '/'), '/'));
+
+        if ($directories) {
+            $from .= '/dummy_file';
+        }
+
         if (dirname($from) === dirname($to)) {
             return './'.basename($to);
         }
-        $from = lcfirst(rtrim(strtr($from, '\\', '/'), '/'));
-        $to = lcfirst(rtrim(strtr($to, '\\', '/'), '/'));
 
         $commonPath = $to;
         while (strpos($from, $commonPath) !== 0 && '/' !== $commonPath && !preg_match('{^[a-z]:/?$}i', $commonPath) && '.' !== $commonPath) {
@@ -94,11 +107,12 @@ class Filesystem
             throw new \InvalidArgumentException('from and to must be absolute paths');
         }
 
+        $from = lcfirst(strtr($from, '\\', '/'));
+        $to = lcfirst(strtr($to, '\\', '/'));
+
         if ($from === $to) {
             return $directories ? '__DIR__' : '__FILE__';
         }
-        $from = lcfirst(strtr($from, '\\', '/'));
-        $to = lcfirst(strtr($to, '\\', '/'));
 
         $commonPath = $to;
         while (strpos($from, $commonPath) !== 0 && '/' !== $commonPath && !preg_match('{^[a-z]:/?$}i', $commonPath) && '.' !== $commonPath) {
